requirements.txt
django>=4.2
psycopg2-binary
Pyfhel
celery~=5.4.0
redis
djangorestframework~=3.15.2
djangorestframework-simplejwt
sentence-transformers
django-celery-results
numpy~=2.0.2